requirements.txt
# AI Financial Statement Fraud Detector - Dependencies
# Compatible with Python 3.9-3.12 (Streamlit Cloud)

streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.18.0
openai>=1.3.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Optional: scikit-learn for Isolation Forest (fallback used if unavailable)
scikit-learn>=1.3.0
